Add user-validations middleware to do the validation of the updated data before going to the controller
